payment_service/pyproject.toml
[tool.poetry]
name = "payment-service"
version = "0.1.0"
description = ""
authors = ["MuhammadHammadBashir0315 <hammad.bashir0315@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlmodel = "^0.0.19"
psycopg2-binary = "^2.9.9"
stripe = "^10.1.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
confluent-kafka = "^2.4.0"
protobuf = "^5.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
